--- api_gateway/src/main/kotlin/org/thechance/api_gateway/endpoints/model/OrderDto.kt
@@ -1,20 +1,0 @@
-package org.thechance.api_gateway.endpoints.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Order(
-    val id: String,
-    val userId: String,
-    val restaurantId: String,
-    val meals: List<Meal>,
-    val totalPrice: Double,
-    val createdAt: Long,
-    val orderStatus: Int = 0
-){
-    @Serializable
-    data class Meal(
-        val mealId: String,
-        val quantity: Int
-    )
-}
